의존성 구조 변경 및 코어 변경

+++ core/lib/domain/course/model/course_score_entity.dart
@@ -4,6 +4,6 @@ part 'course_score_entity.freezed.dart';
 
 @freezed
 abstract class CourseScoreEntity with _$CourseScoreEntity {
-  const factory CourseScoreEntity({required String name, required int score}) =
+  const factory CourseScoreEntity({required String name, required double score}) =
       _CourseScoreEntity;
 }
